[DTR-5216] - Refactor submission and polling mappers
