+++ app/uk/gov/hmrc/constructionindustryscheme/services/chris/ChrisVerificationSubmissionXmlMapper.scala
@@ -18,44 +18,10 @@ package uk.gov.hmrc.constructionindustryscheme.services.chris
 
 import uk.gov.hmrc.constructionindustryscheme.models.*
 
-import scala.xml.*
-
 object ChrisVerificationSubmissionXmlMapper extends ChrisXmlMapper {
 
-  def parse(xml: String): Either[String, SubmissionResult] = {
-    val doc            = XML.loadString(xml)
-    val messageDetails = doc \\ "Header" \\ "MessageDetails"
-
-    for {
-      qualifier                     <- textRequired(messageDetails, "Qualifier", "Qualifier")
-      function                      <- textRequired(messageDetails, "Function", "Function")
-      className                     <- textRequired(messageDetails, "Class", "Class")
-      correlationId                 <- textRequired(messageDetails, "CorrelationID", "CorrelationID")
-      gatewayTimestampOpt            = textOptional(messageDetails, "GatewayTimestamp")
-      acceptedTime                   = textOptional(doc \\ "Body" \ "SuccessResponse", "AcceptedTime")
-      pollIntervalOpt: Option[Int]   = intAttrOptional(messageDetails, "ResponseEndPoint", "PollInterval")
-      endpointUrlOpt: Option[String] = textOptional(messageDetails, "ResponseEndPoint")
-      errOpt                        <- parseError(qualifier, doc)
-    } yield {
-      val status: SubmissionStatus = deriveInitialStatus(qualifier)
-
-      val pollInt = pollIntervalOpt.getOrElse(0)
-      val epUrl   = endpointUrlOpt.getOrElse("")
-
-      val meta = GovTalkMeta(
-        qualifier = qualifier,
-        function = function,
-        className = className,
-        correlationId = correlationId,
-        gatewayTimestamp = gatewayTimestampOpt,
-        responseEndPoint = ResponseEndPoint(epUrl, pollInt),
-        error = errOpt,
-        acceptedTime = acceptedTime
-      )
-
-      SubmissionResult(status, xml, meta)
-    }
-  }
+  def parse(xml: String): Either[String, SubmissionResult] =
+    parseSubmission(xml)(deriveInitialStatus)
 
   private def deriveInitialStatus(qualifier: String): SubmissionStatus =
     qualifier.toLowerCase match {
