+++ app/uk/gov/hmrc/constructionindustryscheme/services/chris/ChrisXmlMapper.scala
@@ -16,8 +16,12 @@
 
 package uk.gov.hmrc.constructionindustryscheme.services.chris
 
-import uk.gov.hmrc.constructionindustryscheme.models.GovTalkError
+import play.api.libs.json.Json
+import uk.gov.hmrc.constructionindustryscheme.models.*
+import uk.gov.hmrc.constructionindustryscheme.models.response.ChrisPollResponse
 
+import java.time.{Instant, LocalDateTime, ZoneId}
+import scala.util.Try
 import scala.xml.*
 
 trait ChrisXmlMapper {
@@ -45,5 +49,102 @@ trait ChrisXmlMapper {
         errorText   <- textRequired(e, "Text", "GovTalkErrors/Error/Text")
       } yield Some(GovTalkError(errorNumber = errorNumber, errorType = errorType, errorText = errorText))
     } else Right(None)
+
+  /** Shared parsing for stage-1 (submit) GovTalk responses. Callers supply the qualifier => status mapping appropriate
+    * to their flow.
+    */
+  protected def parseSubmission(xml: String)(
+    deriveStatus: String => SubmissionStatus
+  ): Either[String, SubmissionResult] = {
+    val doc            = XML.loadString(xml)
+    val messageDetails = doc \\ "Header" \\ "MessageDetails"
+
+    for {
+      qualifier                     <- textRequired(messageDetails, "Qualifier", "Qualifier")
+      function                      <- textRequired(messageDetails, "Function", "Function")
+      className                     <- textRequired(messageDetails, "Class", "Class")
+      correlationId                 <- textRequired(messageDetails, "CorrelationID", "CorrelationID")
+      gatewayTimestampOpt            = textOptional(messageDetails, "GatewayTimestamp")
+      acceptedTime                   = textOptional(doc \\ "Body" \ "SuccessResponse", "AcceptedTime")
+      pollIntervalOpt: Option[Int]   = intAttrOptional(messageDetails, "ResponseEndPoint", "PollInterval")
+      endpointUrlOpt: Option[String] = textOptional(messageDetails, "ResponseEndPoint")
+      errOpt                        <- parseError(qualifier, doc)
+    } yield {
+      val meta = GovTalkMeta(
+        qualifier = qualifier,
+        function = function,
+        className = className,
+        correlationId = correlationId,
+        gatewayTimestamp = gatewayTimestampOpt,
+        responseEndPoint = ResponseEndPoint(endpointUrlOpt.getOrElse(""), pollIntervalOpt.getOrElse(0)),
+        error = errOpt,
+        acceptedTime = acceptedTime
+      )
+
+      SubmissionResult(deriveStatus(qualifier), xml, meta)
+    }
+  }
+
+  /** Shared parsing for stage-2 (poll) GovTalk responses. Callers supply the status mapping appropriate to their flow.
+    */
+  protected def parsePoll(xml: String, now: Instant)(
+    deriveStatus: (String, Option[GovTalkError], Elem) => SubmissionStatus
+  ): Either[String, ChrisPollResponse] = {
+    val doc            = XML.loadString(xml)
+    val messageDetails = doc \\ "Header" \\ "MessageDetails"
+
+    for {
+      qualifier                     <- textRequired(messageDetails, "Qualifier", "Qualifier")
+      correlationId                 <- textRequired(messageDetails, "CorrelationID", "CorrelationID")
+      endpointUrlOpt: Option[String] = textOptional(messageDetails, "ResponseEndPoint")
+      pollIntervalOpt: Option[Int]   = intAttrOptional(messageDetails, "ResponseEndPoint", "PollInterval")
+      lastMessageDateOpt            <- gatewayTimeStampOrNow(messageDetails, now)
+      acceptedTime                   = textOptional(doc \\ "Body" \ "SuccessResponse", "AcceptedTime")
+      errOpt                        <- parseError(qualifier, doc)
+      irMark                         = textOptional(
+                                         doc \\ "Body" \ "SuccessResponse" \ "IRmarkReceipt" \ "Signature" \ "SignedInfo" \ "Reference",
+                                         "DigestValue"
+                                       )
+    } yield ChrisPollResponse(
+      deriveStatus(qualifier, errOpt, doc),
+      correlationId,
+      endpointUrlOpt,
+      pollIntervalOpt,
+      errOpt.map(Json.toJson(_)),
+      irMark,
+      lastMessageDateOpt,
+      acceptedTime
+    )
+  }
+
+  /** Detects an IRMark mismatch error inside the <Body> ErrorResponse. */
+  protected def isIrmarkMismatch(doc: Elem): Boolean = {
+    val bodyErrors = (doc \\ "Body") \\ "Error"
+
+    bodyErrors.exists { e =>
+      val number    = (e \ "Number").text.trim
+      val errorType = (e \ "Type").text.trim
+      val text      = (e \ "Text").text.trim.toLowerCase
+
+      number == "2021" &&
+      errorType.equalsIgnoreCase("business") &&
+      text.contains("irmark") &&
+      text.contains("incorrect")
+    }
+  }
+
+  private val UkZone = ZoneId.of("Europe/London")
+
+  private def gatewayTimeStampOrNow(messageDetails: NodeSeq, now: Instant): Either[String, Option[String]] =
+    textOptional(messageDetails, "GatewayTimestamp") match {
+      case Some(raw) => normaliseGatewayTimestamp(raw).map(ts => Some(ts): Option[String])
+      case None      => Right(Some(now.toString): Option[String])
+    }
+
+  private def normaliseGatewayTimestamp(raw: String): Either[String, String] =
+    Try(LocalDateTime.parse(raw).atZone(UkZone).toInstant).toEither match {
+      case Right(value) => Right(value.toString)
+      case Left(err)    => Left(s"Failed to parse GatewayTimestamp '$raw': ${err.getMessage}")
+    }
 
 }
